apply text to both cases, and update tests
